+++ packages/connect-kit/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Don't open a blank tab when pressing the User Ledger Live deeplink.
+- Correctly detect the Brave browser on iOS.
 - Patch the provider's request method to handle the Connect With Ledger Live
   modal in the `eth_requestAccounts` method, allowing us to reject with the
   "user rejected connection" error when the modal is closed.